+++ cpp/benchmarks/merge/merge_benchmark.cpp
@@ -79,7 +79,7 @@ void BM_merge(benchmark::State& state)
 
   for (auto _ : state) {
     cuda_event_timer raii(state, true);  // flush_l2_cache = true, stream = 0
-    auto result = cudf::experimental::merge(tables, key_cols, column_order, null_precedence);
+    auto result = cudf::merge(tables, key_cols, column_order, null_precedence);
   }
 
   state.SetBytesProcessed(state.iterations() * 2 * sizeof(int32_t) * total_rows);