+++ cpp/include/cudf/merge.hpp
@@ -16,12 +16,11 @@
 
 #pragma once
 
-#include <cudf/cudf.h>
 #include <cudf/types.hpp>
+#include <memory>
 #include <vector>
 
 namespace cudf {
-namespace experimental {
 /**
  * @addtogroup column_merge
  * @{
@@ -93,7 +92,7 @@ namespace experimental {
  *
  * @returns A table containing sorted data from all input tables
  */
-std::unique_ptr<cudf::experimental::table> merge(
+std::unique_ptr<cudf::table> merge(
   std::vector<table_view> const& tables_to_merge,
   std::vector<cudf::size_type> const& key_cols,
   std::vector<cudf::order> const& column_order,
@@ -101,5 +100,4 @@ std::unique_ptr<cudf::experimental::table> merge(
   rmm::mr::device_memory_resource* mr                  = rmm::mr::get_default_resource());
 
 /** @} */  // end of group
-}  // namespace experimental
 }  // namespace cudf