+++ cpp/include/cudf/utilities/type_dispatcher.hpp
@@ -29,7 +29,6 @@
  * and concrete C++ types.
  **/
 namespace cudf {
-namespace experimental {
 /**
  * @addtogroup utility_dispatcher
  * @{
@@ -372,5 +371,4 @@ CUDA_HOST_DEVICE_CALLABLE constexpr decltype(auto) type_dispatcher(cudf::data_ty
 }
 
 /** @} */  // end of group
-}  // namespace experimental
 }  // namespace cudf