+++ cpp/src/binaryop/jit/code/traits.cpp
@@ -18,7 +18,6 @@
  */
 
 namespace cudf {
-namespace experimental {
 namespace binops {
 namespace jit {
 namespace code {
@@ -63,5 +62,4 @@ const char* traits =
 }  // namespace code
 }  // namespace jit
 }  // namespace binops
-}  // namespace experimental
 }  // namespace cudf