+++ cpp/src/strings/split/split_record.cu
@@ -36,9 +36,9 @@ static constexpr size_type split_align = 64;
 
 __device__ size_type compute_memory_size(size_type token_count, size_type token_size_sum)
 {
-  return cudf::experimental::detail::round_up_pow2(token_size_sum, split_align) +
-         cudf::experimental::detail::round_up_pow2(
-           (token_count + 1) * static_cast<size_type>(sizeof(size_type)), split_align);
+  return cudf::detail::round_up_pow2(token_size_sum, split_align) +
+         cudf::detail::round_up_pow2((token_count + 1) * static_cast<size_type>(sizeof(size_type)),
+                                     split_align);
 }
 
 struct copy_info {
@@ -152,9 +152,8 @@ struct token_copier_fn {
 
     auto memory_ptr = static_cast<char*>(info.memory_ptr);
 
-    auto const char_buf_size =
-      cudf::experimental::detail::round_up_pow2(info.token_size_sum, split_align);
-    auto const char_buf_ptr = memory_ptr;
+    auto const char_buf_size = cudf::detail::round_up_pow2(info.token_size_sum, split_align);
+    auto const char_buf_ptr  = memory_ptr;
     memory_ptr += char_buf_size;
     auto const offset_buf_ptr = reinterpret_cast<size_type*>(memory_ptr);
 
@@ -324,9 +323,8 @@ struct whitespace_token_copier_fn {
 
     auto memory_ptr = static_cast<char*>(info.memory_ptr);
 
-    auto const char_buf_size =
-      cudf::experimental::detail::round_up_pow2(info.token_size_sum, split_align);
-    auto const char_buf_ptr = memory_ptr;
+    auto const char_buf_size = cudf::detail::round_up_pow2(info.token_size_sum, split_align);
+    auto const char_buf_ptr  = memory_ptr;
     memory_ptr += char_buf_size;
     auto const offset_buf_ptr = reinterpret_cast<size_type*>(memory_ptr);