+++ cpp/tests/groupby/group_argmin_test.cpp
@@ -35,7 +35,7 @@ TYPED_TEST(groupby_argmin_test, basic)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::ARGMIN>;
+    using R = cudf::detail::target_type_t<V, aggregation::ARGMIN>;
 
     if (std::is_same<V, bool>::value) return;
 
@@ -45,10 +45,10 @@ TYPED_TEST(groupby_argmin_test, basic)
     fixed_width_column_wrapper<K> expect_keys { 1, 2, 3 };
     fixed_width_column_wrapper<R> expect_vals { 6, 9, 8 };
 
-    auto agg = cudf::experimental::make_argmin_aggregation();
+    auto agg = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg2 = cudf::experimental::make_argmin_aggregation();
+    auto agg2 = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2), force_use_sort_impl::YES);
 }
 
@@ -56,7 +56,7 @@ TYPED_TEST(groupby_argmin_test, zero_valid_keys)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::ARGMIN>;
+    using R = cudf::detail::target_type_t<V, aggregation::ARGMIN>;
 
     if (std::is_same<V, bool>::value) return;
 
@@ -66,10 +66,10 @@ TYPED_TEST(groupby_argmin_test, zero_valid_keys)
     fixed_width_column_wrapper<K> expect_keys { };
     fixed_width_column_wrapper<R> expect_vals { };
 
-    auto agg = cudf::experimental::make_argmin_aggregation();
+    auto agg = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg2 = cudf::experimental::make_argmin_aggregation();
+    auto agg2 = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2), force_use_sort_impl::YES);
 }
 
@@ -77,7 +77,7 @@ TYPED_TEST(groupby_argmin_test, zero_valid_values)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::ARGMIN>;
+    using R = cudf::detail::target_type_t<V, aggregation::ARGMIN>;
 
     if (std::is_same<V, bool>::value) return;
 
@@ -87,10 +87,10 @@ TYPED_TEST(groupby_argmin_test, zero_valid_values)
     fixed_width_column_wrapper<K> expect_keys { 1 };
     fixed_width_column_wrapper<R> expect_vals({ 0 }, all_null());
 
-    auto agg = cudf::experimental::make_argmin_aggregation();
+    auto agg = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg2 = cudf::experimental::make_argmin_aggregation();
+    auto agg2 = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2), force_use_sort_impl::YES);
 }
 
@@ -98,7 +98,7 @@ TYPED_TEST(groupby_argmin_test, null_keys_and_values)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::ARGMIN>;
+    using R = cudf::detail::target_type_t<V, aggregation::ARGMIN>;
 
     if (std::is_same<V, bool>::value) return;
 
@@ -113,11 +113,11 @@ TYPED_TEST(groupby_argmin_test, null_keys_and_values)
     fixed_width_column_wrapper<R> expect_vals({ 3,        9,         8,       0},
                                               { 1,        1,         1,       0});
 
-    auto agg = cudf::experimental::make_argmin_aggregation();
+    auto agg = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
     // TODO: explore making this a gtest parameter
-    auto agg2 = cudf::experimental::make_argmin_aggregation();
+    auto agg2 = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2), force_use_sort_impl::YES);
 }
 
@@ -128,7 +128,7 @@ TEST_F(groupby_argmin_string_test, basic)
 {
     using K = int32_t;
     using V = string_view;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::ARGMIN>;
+    using R = cudf::detail::target_type_t<V, aggregation::ARGMIN>;
 
     fixed_width_column_wrapper<K> keys        {     1,     2,    3,     1,     2,     2,     1,    3,    3,    2 };
     strings_column_wrapper        vals        { "año", "bit", "₹1", "aaa", "zit", "bat", "aab", "$1", "€1", "wut"};
@@ -136,10 +136,10 @@ TEST_F(groupby_argmin_string_test, basic)
     fixed_width_column_wrapper<K> expect_keys { 1, 2, 3 };
     fixed_width_column_wrapper<R> expect_vals({ 3, 5, 7 });
 
-    auto agg = cudf::experimental::make_argmin_aggregation();
+    auto agg = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg2 = cudf::experimental::make_argmin_aggregation();
+    auto agg2 = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2), force_use_sort_impl::YES);
 }
 
@@ -147,7 +147,7 @@ TEST_F(groupby_argmin_string_test, zero_valid_values)
 {
     using K = int32_t;
     using V = string_view;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::ARGMIN>;
+    using R = cudf::detail::target_type_t<V, aggregation::ARGMIN>;
 
     fixed_width_column_wrapper<K> keys        { 1, 1, 1};
     strings_column_wrapper        vals      ( { "año", "bit", "₹1"}, all_null() );
@@ -155,10 +155,10 @@ TEST_F(groupby_argmin_string_test, zero_valid_values)
     fixed_width_column_wrapper<K> expect_keys { 1 };
     fixed_width_column_wrapper<R> expect_vals({ 0 }, all_null());
 
-    auto agg = cudf::experimental::make_argmin_aggregation();
+    auto agg = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 
-    auto agg2 = cudf::experimental::make_argmin_aggregation();
+    auto agg2 = cudf::make_argmin_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg2), force_use_sort_impl::YES);
 }
 // clang-format on