+++ cpp/tests/groupby/group_mean_test.cpp
@@ -37,7 +37,7 @@ TYPED_TEST(groupby_mean_test, basic)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::MEAN>;
+    using R = cudf::detail::target_type_t<V, aggregation::MEAN>;
 
     fixed_width_column_wrapper<K> keys        { 1, 2, 3, 1, 2, 2, 1, 3, 3, 2};
     fixed_width_column_wrapper<V> vals        { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
@@ -45,7 +45,7 @@ TYPED_TEST(groupby_mean_test, basic)
     fixed_width_column_wrapper<K> expect_keys { 1,  2,     3    };
     fixed_width_column_wrapper<R> expect_vals { 3., 19./4, 17./3};
 
-    auto agg = cudf::experimental::make_mean_aggregation();
+    auto agg = cudf::make_mean_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 }
 
@@ -53,7 +53,7 @@ TYPED_TEST(groupby_mean_test, empty_cols)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::MEAN>;
+    using R = cudf::detail::target_type_t<V, aggregation::MEAN>;
 
     fixed_width_column_wrapper<K> keys        { };
     fixed_width_column_wrapper<V> vals        { };
@@ -61,7 +61,7 @@ TYPED_TEST(groupby_mean_test, empty_cols)
     fixed_width_column_wrapper<K> expect_keys { };
     fixed_width_column_wrapper<R> expect_vals { };
 
-    auto agg = cudf::experimental::make_mean_aggregation();
+    auto agg = cudf::make_mean_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 }
 
@@ -69,7 +69,7 @@ TYPED_TEST(groupby_mean_test, zero_valid_keys)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::MEAN>;
+    using R = cudf::detail::target_type_t<V, aggregation::MEAN>;
 
     fixed_width_column_wrapper<K> keys      ( { 1, 2, 3}, all_null() );
     fixed_width_column_wrapper<V> vals        { 3, 4, 5};
@@ -77,7 +77,7 @@ TYPED_TEST(groupby_mean_test, zero_valid_keys)
     fixed_width_column_wrapper<K> expect_keys { };
     fixed_width_column_wrapper<R> expect_vals { };
 
-    auto agg = cudf::experimental::make_mean_aggregation();
+    auto agg = cudf::make_mean_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 }
 
@@ -85,7 +85,7 @@ TYPED_TEST(groupby_mean_test, zero_valid_values)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::MEAN>;
+    using R = cudf::detail::target_type_t<V, aggregation::MEAN>;
 
     fixed_width_column_wrapper<K> keys        { 1, 1, 1};
     fixed_width_column_wrapper<V> vals      ( { 3, 4, 5}, all_null() );
@@ -93,7 +93,7 @@ TYPED_TEST(groupby_mean_test, zero_valid_values)
     fixed_width_column_wrapper<K> expect_keys { 1 };
     fixed_width_column_wrapper<R> expect_vals({ 0 }, all_null());
 
-    auto agg = cudf::experimental::make_mean_aggregation();
+    auto agg = cudf::make_mean_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 }
 
@@ -101,7 +101,7 @@ TYPED_TEST(groupby_mean_test, null_keys_and_values)
 {
     using K = int32_t;
     using V = TypeParam;
-    using R = experimental::detail::target_type_t<V, experimental::aggregation::MEAN>;
+    using R = cudf::detail::target_type_t<V, aggregation::MEAN>;
 
     fixed_width_column_wrapper<K> keys(       { 1, 2, 3, 1, 2, 2, 1, 3, 3, 2, 4},
                                               { 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1});
@@ -114,7 +114,7 @@ TYPED_TEST(groupby_mean_test, null_keys_and_values)
     fixed_width_column_wrapper<R> expect_vals({ 4.5,      14./3,     5.,      0.},
                                               { 1,        1,         1,       0});
 
-    auto agg = cudf::experimental::make_mean_aggregation();
+    auto agg = cudf::make_mean_aggregation();
     test_single_agg(keys, vals, expect_keys, expect_vals, std::move(agg));
 }
 // clang-format on