+++ cpp/tests/quantiles/quantiles_test.cu
@@ -44,7 +44,7 @@ TYPED_TEST(QuantilesTest, TestZeroColumns)
 {
   auto input = table_view(std::vector<column_view>{});
 
-  EXPECT_THROW(experimental::quantiles(input, {0.0f}), logic_error);
+  EXPECT_THROW(quantiles(input, {0.0f}), logic_error);
 }
 
 TYPED_TEST(QuantilesTest, TestMultiColumnZeroRows)
@@ -54,7 +54,7 @@ TYPED_TEST(QuantilesTest, TestMultiColumnZeroRows)
   auto input_a = fixed_width_column_wrapper<T>({});
   auto input   = table_view({input_a});
 
-  EXPECT_THROW(experimental::quantiles(input, {0.0f}), logic_error);
+  EXPECT_THROW(quantiles(input, {0.0f}), logic_error);
 }
 
 TYPED_TEST(QuantilesTest, TestZeroRequestedQuantiles)
@@ -64,8 +64,8 @@ TYPED_TEST(QuantilesTest, TestZeroRequestedQuantiles)
   auto input_a = fixed_width_column_wrapper<T>({1}, {1});
   auto input   = table_view(std::vector<column_view>{input_a});
 
-  auto actual   = experimental::quantiles(input, {});
-  auto expected = experimental::empty_like(input);
+  auto actual   = quantiles(input, {});
+  auto expected = empty_like(input);
 
   expect_tables_equal(expected->view(), actual->view());
 }
@@ -78,12 +78,12 @@ TYPED_TEST(QuantilesTest, TestMultiColumnOrderCountMismatch)
   auto input_b = fixed_width_column_wrapper<T>({});
   auto input   = table_view({input_a});
 
-  EXPECT_THROW(experimental::quantiles(input,
-                                       {0.0f},
-                                       experimental::interpolation::NEAREST,
-                                       sorted::NO,
-                                       {order::ASCENDING},
-                                       {null_order::AFTER, null_order::AFTER}),
+  EXPECT_THROW(quantiles(input,
+                         {0.0f},
+                         interpolation::NEAREST,
+                         sorted::NO,
+                         {order::ASCENDING},
+                         {null_order::AFTER, null_order::AFTER}),
                logic_error);
 }
 
@@ -95,12 +95,12 @@ TYPED_TEST(QuantilesTest, TestMultiColumnNullOrderCountMismatch)
   auto input_b = fixed_width_column_wrapper<T>({});
   auto input   = table_view({input_a});
 
-  EXPECT_THROW(experimental::quantiles(input,
-                                       {0.0f},
-                                       experimental::interpolation::NEAREST,
-                                       sorted::NO,
-                                       {order::ASCENDING, order::ASCENDING},
-                                       {null_order::AFTER}),
+  EXPECT_THROW(quantiles(input,
+                         {0.0f},
+                         interpolation::NEAREST,
+                         sorted::NO,
+                         {order::ASCENDING, order::ASCENDING},
+                         {null_order::AFTER}),
                logic_error);
 }
 
@@ -112,11 +112,9 @@ TYPED_TEST(QuantilesTest, TestMultiColumnArithmeticInterpolation)
   auto input_b = fixed_width_column_wrapper<T>({});
   auto input   = table_view({input_a});
 
-  EXPECT_THROW(experimental::quantiles(input, {0.0f}, experimental::interpolation::LINEAR),
-               logic_error);
+  EXPECT_THROW(quantiles(input, {0.0f}, interpolation::LINEAR), logic_error);
 
-  EXPECT_THROW(experimental::quantiles(input, {0.0f}, experimental::interpolation::MIDPOINT),
-               logic_error);
+  EXPECT_THROW(quantiles(input, {0.0f}, interpolation::MIDPOINT), logic_error);
 }
 
 TYPED_TEST(QuantilesTest, TestMultiColumnUnsorted)
@@ -134,11 +132,11 @@ TYPED_TEST(QuantilesTest, TestMultiColumnUnsorted)
 
   auto input = table_view({input_a, input_b});
 
-  auto actual = experimental::quantiles(input,
-                                        {0.0f, 0.5f, 0.7f, 0.25f, 1.0f},
-                                        experimental::interpolation::NEAREST,
-                                        sorted::NO,
-                                        {order::ASCENDING, order::DESCENDING});
+  auto actual = quantiles(input,
+                          {0.0f, 0.5f, 0.7f, 0.25f, 1.0f},
+                          interpolation::NEAREST,
+                          sorted::NO,
+                          {order::ASCENDING, order::DESCENDING});
 
   auto expected_a = strings_column_wrapper({"A", "C", "C", "B", "D"}, {1, 1, 1, 1, 1});
 
@@ -164,8 +162,8 @@ TYPED_TEST(QuantilesTest, TestMultiColumnAssumedSorted)
 
   auto input = table_view({input_a, input_b});
 
-  auto actual = experimental::quantiles(
-    input, {0.0f, 0.5f, 0.7f, 0.25f, 1.0f}, experimental::interpolation::NEAREST, sorted::YES);
+  auto actual =
+    quantiles(input, {0.0f, 0.5f, 0.7f, 0.25f, 1.0f}, interpolation::NEAREST, sorted::YES);
 
   auto expected_a = strings_column_wrapper({"C", "D", "C", "D", "A"}, {1, 1, 1, 1, 1});