+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# cuDF 0.15.0 (Date TBD)
+
+## New Features
+
+## Improvements
+
+## Bug Fixes
+
 # cuDF 0.14.0 (Date TBD)
 
 ## New Features
@@ -58,6 +66,7 @@
 - PR #4548 Remove string_view is_null method
 - PR #4645 Add Alias for `kurtosis` as `kurt`
 - PR #4703 Optimize strings concatenate for many columns
+- PR #4769 Remove legacy code from libcudf
 - PR #4668 Add Java bindings for log2/log10 unary ops and log_base binary op
 - PR #4616 Enable different RMM allocation modes in unit tests
 - PR #4520 Fix several single char -> single char case mapping values. Add support for single -> multi char mappings.
@@ -191,6 +200,8 @@
 - PR #5208 Port search and join benchmark to libcudf++
 - PR #5214 Move docs build script into repository
 - PR #5219 Add per context cache for JIT kernels
+- PR #5233 Remove experimental namespace used during libcudf++ refactor
+- PR #5251 Fix more mispellings in cpp comments and strings
 
 ## Bug Fixes
 
@@ -316,6 +327,7 @@
 - PR #5212 Fix memory leak in `dlpack.pyx:from_dlpack()`
 - PR #5224 Add new headers from 5198 to libcudf/meta.yaml
 - PR #5228 Fix datetime64 scalar dtype handling for unsupported time units
+- PR #5256 ORC reader: fix loading individual timestamp columns
 
 
 # cuDF 0.13.0 (31 Mar 2020)